refactor: useTransactionToast hook for improved toast handling
